% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{fetch_redcap_instruments}
\alias{fetch_redcap_instruments}
\title{Fetch REDCap instrument names from a project}
\usage{
fetch_redcap_instruments(con)
}
\arguments{
\item{con}{A \code{redcapAPI::redcapConnection} object.}
}
\value{
A data frame with one column, \code{name}, containing the instrument/form names.
}
\description{
Queries the REDCap project for its list of instruments (forms) and returns
a data frame with a column \code{name} that can be used for dynamic branching in
\code{targets} pipelines.
}
\details{
This function is used internally by \code{\link[=tar_redcap]{tar_redcap()}} to determine which
instruments to pull.
}
\keyword{internal}
