% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{validate_fetch_records}
\alias{validate_fetch_records}
\title{Validate that the \code{fetch_records} function has required arguments}
\usage{
validate_fetch_records(fetch_records)
}
\arguments{
\item{fetch_records}{A function meant to be passed to the \code{fetch_records}
argument of \code{\link[=tar_redcap]{tar_redcap()}}. It must accept at least \code{con} and \code{forms}.}
}
\value{
Invisibly returns \code{TRUE} if validation passes; otherwise an
informative error is thrown.
}
\description{
Ensures that the user-supplied \code{fetch_records} function includes the
required arguments \code{con} and \code{forms}. Helps avoid runtime errors when
dynamic targets call the function inside \code{\link[=tar_redcap]{tar_redcap()}}.
}
\keyword{internal}
