% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{validate_redcap_connection}
\alias{validate_redcap_connection}
\title{Validate REDCap connection object}
\usage{
validate_redcap_connection(con)
}
\arguments{
\item{con}{Object to validate.}
}
\description{
Ensures the object is a redcapAPI connection.
}
\keyword{internal}
